feat(assets): Copy any assets to target folder
With this commit, we introduce a new feature that we copy any assets
that exists in the same source folder than the markdown file and save it
in the target folder.

With this you can reference images or any file and we take care that
those files are available on the generated static folder.